Added full fledged Spectrum class to replace Vector3f typedef - currently only rgb spectrum, but hidden away under a facade, so should be simple to move to real sampled spectrum - resolves #10

+++ include/triangle.h
@@ -112,7 +112,7 @@ public:
 
 	inline int32_t triangleCount() const
 	{
-		return triangles_.size();
+		return static_cast<int32_t>(triangles_.size());
 	}
 
     inline Bsdf* getBsdf() const
